Add cobia reporting exporter migration to dragonfish

+++ src/middlewared/middlewared/alembic/versions/24.04/2023-09-06_14-45_cleanup_reporting_api.py
@@ -7,6 +7,7 @@ Create Date: 2023-09-06 14:45:13.261715+00:00
 
 """
 from alembic import op
+from sqlalchemy.engine.reflection import Inspector
 
 
 revision = 'e915a3b8fff6'
@@ -16,6 +17,11 @@ depends_on = None
 
 
 def upgrade():
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    if 'system_reporting' not in inspector.get_table_names():
+        return  # Skip as backported cobia migration already removed system_reporting table
+
     with op.batch_alter_table('system_reporting', schema=None) as batch_op:
         batch_op.drop_column('graph_age')
         batch_op.drop_column('graph_points')
